rules: move assertions and predicates
This commit refactors assertions and predicates in the rules package.

Release note: None
