+++ pkg/sql/schemachanger/scplan/internal/rules/helpers.go
@@ -15,11 +15,14 @@ import (
 	"reflect"
 	"strings"
 
+	"github.com/cockroachdb/cockroach/pkg/sql/catalog/catpb"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descpb"
 	"github.com/cockroachdb/cockroach/pkg/sql/schemachanger/rel"
 	"github.com/cockroachdb/cockroach/pkg/sql/schemachanger/scpb"
+	"github.com/cockroachdb/cockroach/pkg/sql/schemachanger/scplan/internal/opgen"
 	"github.com/cockroachdb/cockroach/pkg/sql/schemachanger/screl"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/catid"
+	"github.com/cockroachdb/cockroach/pkg/sql/types"
 	"github.com/cockroachdb/cockroach/pkg/util/iterutil"
 	"github.com/cockroachdb/errors"
 )
@@ -202,12 +205,20 @@ func forEachElement(fn func(element scpb.Element) error) error {
 	return nil
 }
 
-func elementTypes(nv nodeVars, filter func(element scpb.Element) bool) rel.Clause {
+// elementTypes returns a Type clause which binds the element var to elements of
+// a specific type, filtered by the conjunction of all provided predicates.
+func elementTypes(nv nodeVars, filters ...func(element scpb.Element) bool) rel.Clause {
+	if len(filters) == 0 {
+		panic(errors.AssertionFailedf("empty filter predicate list for var %q", nv))
+	}
 	var types []interface{}
 	_ = forEachElement(func(e scpb.Element) error {
-		if filter(e) {
-			types = append(types, e)
+		for _, filter := range filters {
+			if !filter(e) {
+				return nil
+			}
 		}
+		types = append(types, e)
 		return nil
 	})
 	if len(types) == 0 {
@@ -218,4 +229,171 @@ func elementTypes(nv nodeVars, filter func(element scpb.Element) bool) rel.Claus
 
 func nonNilElement(element scpb.Element) scpb.Element {
 	return reflect.New(reflect.ValueOf(element).Type().Elem()).Interface().(scpb.Element)
+}
+
+// IsDescriptor returns true for a descriptor-element, i.e. an element which
+// owns its corresponding descriptor.
+func IsDescriptor(e scpb.Element) bool {
+	switch e.(type) {
+	case *scpb.Database, *scpb.Schema, *scpb.Table, *scpb.View, *scpb.Sequence, *scpb.AliasType, *scpb.EnumType:
+		return true
+	}
+	return false
+}
+
+func isSubjectTo2VersionInvariant(e scpb.Element) bool {
+	switch e.(type) {
+	case *scpb.Column, *scpb.PrimaryIndex, *scpb.SecondaryIndex, *scpb.TemporaryIndex:
+		return true
+	}
+	return false
+}
+
+func isSimpleDependent(e scpb.Element) bool {
+	return !IsDescriptor(e) && !isSubjectTo2VersionInvariant(e)
+}
+
+// Assert that only simple dependents (non-descriptor, non-index, non-column)
+// have screl.ReferencedDescID attributes.
+func init() {
+	_ = forEachElement(func(e scpb.Element) error {
+		if isSimpleDependent(e) {
+			return nil
+		}
+		e = nonNilElement(e)
+		if _, err := screl.Schema.GetAttribute(screl.ReferencedDescID, e); err == nil {
+			panic(errors.AssertionFailedf("%T not expected to have screl.ReferencedDescID attr", e))
+		}
+		return nil
+	})
+}
+
+// Assert that elements can be grouped into three categories when transitioning
+// from PUBLIC to ABSENT:
+// - go via DROPPED iff they're descriptor elements
+// - go via a non-read status iff they're indexes or columns, which are
+//   subject to the two-version invariant.
+// - go direct to ABSENT in all other cases.
+func init() {
+	_ = forEachElement(func(e scpb.Element) error {
+		s0 := opgen.InitialStatus(e, scpb.Status_ABSENT)
+		s1 := opgen.NextStatus(e, scpb.Status_ABSENT, s0)
+		switch s1 {
+		case scpb.Status_OFFLINE, scpb.Status_DROPPED:
+			if IsDescriptor(e) {
+				return nil
+			}
+		case scpb.Status_VALIDATED, scpb.Status_WRITE_ONLY, scpb.Status_DELETE_ONLY:
+			if isSubjectTo2VersionInvariant(e) {
+				return nil
+			}
+		case scpb.Status_ABSENT:
+			if isSimpleDependent(e) {
+				return nil
+			}
+		}
+		panic(errors.AssertionFailedf(
+			"unexpected transition %s -> %s in direction ABSENT for %T (descriptor=%v, 2VI=%v)",
+			s0, s1, e, IsDescriptor(e), isSubjectTo2VersionInvariant(e),
+		))
+	})
+}
+
+func getTypeT(element scpb.Element) (*scpb.TypeT, error) {
+	switch e := element.(type) {
+	case *scpb.ColumnType:
+		if e == nil {
+			return nil, nil
+		}
+		return &e.TypeT, nil
+	case *scpb.AliasType:
+		if e == nil {
+			return nil, nil
+		}
+		return &e.TypeT, nil
+	}
+	return nil, errors.AssertionFailedf("element %T does not have an embedded scpb.TypeT", element)
+}
+
+func isWithTypeT(element scpb.Element) bool {
+	_, err := getTypeT(element)
+	return err == nil
+}
+
+func getTypeTOrPanic(element scpb.Element) *scpb.TypeT {
+	ret, err := getTypeT(element)
+	if err != nil {
+		panic(err)
+	}
+	return ret
+}
+
+// Assert that isWithTypeT covers all elements with embedded TypeTs.
+func init() {
+	_ = forEachElement(func(e scpb.Element) error {
+		e = nonNilElement(e)
+		return screl.WalkTypes(e, func(t *types.T) error {
+			if isWithTypeT(e) {
+				return nil
+			}
+			panic(errors.AssertionFailedf("getTypeT should support %T but doesn't", e))
+		})
+	})
+}
+
+func getExpression(element scpb.Element) (*scpb.Expression, error) {
+	switch e := element.(type) {
+	case *scpb.ColumnType:
+		if e == nil {
+			return nil, nil
+		}
+		return e.ComputeExpr, nil
+	case *scpb.ColumnDefaultExpression:
+		if e == nil {
+			return nil, nil
+		}
+		return &e.Expression, nil
+	case *scpb.ColumnOnUpdateExpression:
+		if e == nil {
+			return nil, nil
+		}
+		return &e.Expression, nil
+	case *scpb.SecondaryIndexPartial:
+		if e == nil {
+			return nil, nil
+		}
+		return &e.Expression, nil
+	case *scpb.CheckConstraint:
+		if e == nil {
+			return nil, nil
+		}
+		return &e.Expression, nil
+	}
+	return nil, errors.AssertionFailedf("element %T does not have an embedded scpb.Expression", element)
+}
+
+func isWithExpression(element scpb.Element) bool {
+	_, err := getExpression(element)
+	return err == nil
+}
+
+func getExpressionOrPanic(element scpb.Element) *scpb.Expression {
+	ret, err := getExpression(element)
+	if err != nil {
+		panic(err)
+	}
+	return ret
+}
+
+// Assert that isWithExpression covers all elements with embedded
+// expressions.
+func init() {
+	_ = forEachElement(func(e scpb.Element) error {
+		return screl.WalkExpressions(e, func(t *catpb.Expression) error {
+			if isWithExpression(e) {
+				return nil
+			}
+			panic(errors.AssertionFailedf("getExpression should support %T but doesn't", e))
+		})
+	})
 }
